+++ api/src/main/java/ca/bc/gov/educ/api/assessment/config/EducGradAssessmentApiConfig.java
@@ -1,0 +1,34 @@
+package ca.bc.gov.educ.api.assessment.config;
+
+import ca.bc.gov.educ.api.assessment.model.dto.Assessment;
+import ca.bc.gov.educ.api.assessment.model.entity.AssessmentEntity;
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class EducGradAssessmentApiConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+
+        ModelMapper modelMapper = new ModelMapper();
+        modelMapper.typeMap(AssessmentEntity.class, Assessment.class);
+        modelMapper.typeMap(Assessment.class, AssessmentEntity.class);
+        return modelMapper;
+    }
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.create();
+    }
+
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+
+}
